+++ std/src/exec.rs
@@ -13,7 +13,7 @@
 // License for the specific language governing permissions and limitations
 // under the License.
 
-//! Commands that directly manipulate the machine's state.
+//! Commands that manipulate the machine's state or the program's execution.
 
 use async_trait::async_trait;
 use endbasic_core::ast::{ArgSep, Expr, Value, VarType};
@@ -21,7 +21,13 @@ use endbasic_core::exec::Machine;
 use endbasic_core::syms::{
     CallError, CallableMetadata, CallableMetadataBuilder, Command, CommandResult,
 };
+use futures_lite::future::{BoxedLocal, FutureExt};
 use std::rc::Rc;
+use std::thread;
+use std::time::Duration;
+
+/// Category string for all functions and commands provided by this module.
+const CATEGORY: &str = "Interpreter manipulation";
 
 /// The `CLEAR` command.
 pub struct ClearCommand {
@@ -34,7 +40,7 @@ impl ClearCommand {
         Rc::from(Self {
             metadata: CallableMetadataBuilder::new("CLEAR", VarType::Void)
                 .with_syntax("")
-                .with_category("Interpreter manipulation")
+                .with_category(CATEGORY)
                 .with_description("Clears all variables to restore initial state.")
                 .build(),
         })
@@ -67,7 +73,7 @@ impl ExitCommand {
         Rc::from(Self {
             metadata: CallableMetadataBuilder::new("EXIT", VarType::Void)
                 .with_syntax("[code%]")
-                .with_category("Interpreter manipulation")
+                .with_category(CATEGORY)
                 .with_description(
                     "Exits the interpreter.
 The optional code indicates the return value to return to the system.",
@@ -115,16 +121,95 @@ impl Command for ExitCommand {
     }
 }
 
+/// Type of the sleep function used by the `SLEEP` command to actually suspend execution.
+pub type SleepFn = Box<dyn Fn(Duration) -> BoxedLocal<CommandResult>>;
+
+/// An implementation of a `SleepFn` that stops the current thread.
+fn system_sleep(d: Duration) -> BoxedLocal<CommandResult> {
+    async move {
+        thread::sleep(d);
+        Ok(())
+    }
+    .boxed_local()
+}
+
+/// The `SLEEP` command.
+pub struct SleepCommand {
+    metadata: CallableMetadata,
+    sleep_fn: SleepFn,
+}
+
+impl SleepCommand {
+    /// Creates a new instance of the command.
+    pub fn new(sleep_fn: SleepFn) -> Rc<Self> {
+        Rc::from(Self {
+            metadata: CallableMetadataBuilder::new("SLEEP", VarType::Void)
+                .with_syntax("seconds%|seconds#")
+                .with_category(CATEGORY)
+                .with_description(
+                    "Suspends program execution.
+Pauses program execution for the given number of seconds, which can be specified either as an \
+integral or as a floating point number for finer precision.",
+                )
+                .build(),
+            sleep_fn,
+        })
+    }
+}
+
+#[async_trait(?Send)]
+impl Command for SleepCommand {
+    fn metadata(&self) -> &CallableMetadata {
+        &self.metadata
+    }
+
+    async fn exec(&self, args: &[(Option<Expr>, ArgSep)], machine: &mut Machine) -> CommandResult {
+        let duration = match args {
+            [(Some(expr), ArgSep::End)] => match expr.eval(machine.get_symbols())? {
+                Value::Integer(n) => {
+                    if n < 0 {
+                        return Err(CallError::ArgumentError(
+                            "Sleep time must be positive".to_owned(),
+                        ));
+                    }
+                    Duration::from_secs(n as u64)
+                }
+                Value::Double(n) => {
+                    if n < 0.0 {
+                        return Err(CallError::ArgumentError(
+                            "Sleep time must be positive".to_owned(),
+                        ));
+                    }
+                    Duration::from_secs_f64(n)
+                }
+                _ => {
+                    return Err(CallError::ArgumentError(
+                        "Sleep time must be an integer or a double".to_owned(),
+                    ))
+                }
+            },
+            _ => return Err(CallError::ArgumentError("SLEEP takes one argument".to_owned())),
+        };
+        (self.sleep_fn)(duration).await
+    }
+}
+
 /// Instantiates all REPL commands and adds them to the `machine`.
-pub fn add_all(machine: &mut Machine) {
+///
+/// `sleep_fn` is an async function that implements a pause given a `Duration`.  If not provided,
+/// uses the `std::thread::sleep` function.
+pub fn add_all(machine: &mut Machine, sleep_fn: Option<SleepFn>) {
     machine.add_command(ClearCommand::new());
     machine.add_command(ExitCommand::new());
+    machine.add_command(SleepCommand::new(sleep_fn.unwrap_or_else(|| Box::from(system_sleep))));
 }
 
 #[cfg(test)]
 mod tests {
+    use super::*;
     use crate::testutils::*;
     use endbasic_core::exec::StopReason;
+    use std::time::Instant;
 
     #[test]
     fn test_clear_ok() {
@@ -167,5 +252,36 @@ mod tests {
         check_stmt_err("EXIT takes zero or one argument", "EXIT 1, 2");
         check_stmt_err("Exit code must be a positive integer", "EXIT -3");
         check_stmt_err("Exit code cannot be larger than 127", "EXIT 128");
+    }
+
+    #[test]
+    fn test_sleep_ok() {
+        let sleep_fake = |d: Duration| -> BoxedLocal<CommandResult> {
+            async move { Err(CallError::InternalError(format!("Got {} ms", d.as_millis()))) }
+                .boxed_local()
+        };
+
+        let mut machine = Machine::default();
+        machine.add_command(SleepCommand::new(Box::from(sleep_fake)));
+        let mut t = Tester::from(machine);
+        t.run("SLEEP 123").expect_err("Got 123000 ms").check();
+        t.run("SLEEP 123.1").expect_err("Got 123100 ms").check();
+    }
+
+    #[test]
+    fn test_sleep_real() {
+        let before = Instant::now();
+        Tester::default().run("SLEEP 0.010").check();
+        assert!(before.elapsed() >= Duration::from_millis(10));
+    }
+
+    #[test]
+    fn test_sleep_errors() {
+        check_stmt_err("SLEEP takes one argument", "SLEEP");
+        check_stmt_err("SLEEP takes one argument", "SLEEP 2, 3");
+        check_stmt_err("SLEEP takes one argument", "SLEEP 2; 3");
+        check_stmt_err("Sleep time must be an integer or a double", "SLEEP \"foo\"");
+        check_stmt_err("Sleep time must be positive", "SLEEP -1");
+        check_stmt_err("Sleep time must be positive", "SLEEP -0.001");
     }
 }